emenudsp/requirements.txt
asgiref==3.3.1
colorama==0.4.4
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
qrcode==6.1
six==1.15.0
sqlparse==0.4.1
